+++ bundles/full/src/bundle.ts
@@ -1,0 +1,6 @@
+import { tsParticles } from "tsparticles-engine";
+import { loadFull } from "./index";
+
+loadFull(tsParticles);
+
+export { tsParticles };