+++ bundles/slim/src/bundle.ts
@@ -1,0 +1,6 @@
+import { tsParticles } from "tsparticles-engine";
+import { loadSlim } from "./index";
+
+loadSlim(tsParticles);
+
+export { tsParticles };