+++ engine/src/Types/index.ts
@@ -1,5 +1,5 @@
 export * from "./ParticlesGroups";
-export * from "./Range";
+export * from "./RangeValue";
 export * from "./RecursivePartial";
 export * from "./ShapeData";
 export * from "./ShapeDrawerFunctions";