+++ engine/tests/Utils.ts
@@ -1,3 +1,4 @@
+import { describe, it } from "mocha";
 import { expect } from "chai";
 import type { IContainerPlugin, IPlugin, IRangeValue } from "../src";
 import { Container } from "../src/Core/Container";