+++ plugins/emitters/src/Options/Classes/Emitter.ts
@@ -16,6 +16,7 @@ import { EmitterSpin } from "./EmitterSpin";
  * @category Emitters Plugin
  */
 export class Emitter implements IEmitter, IOptionLoader<IEmitter> {
+    public autoPlay;
     public direction: MoveDirection | keyof typeof MoveDirection | MoveDirectionAlt;
     public life;
     public name?: string;
@@ -27,6 +28,7 @@ export class Emitter implements IEmitter, IOptionLoader<IEmitter> {
     public spawnColor?: AnimatableColor;
 
     constructor() {
+        this.autoPlay = true;
         this.direction = MoveDirection.none;
         this.life = new EmitterLife();
         this.rate = new EmitterRate();
@@ -36,6 +38,10 @@ export class Emitter implements IEmitter, IOptionLoader<IEmitter> {
     public load(data?: RecursivePartial<IEmitter>): void {
         if (data === undefined) {
             return;
+        }
+
+        if (data.autoPlay !== undefined) {
+            this.autoPlay = data.autoPlay;
         }
 
         if (data.size !== undefined) {