+++ demo/inferno/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.0.0-alpha.6](https://github.com/matteobruni/tsparticles/compare/inferno-particles-demo@1.7.2...inferno-particles-demo@2.0.0-alpha.6) (2021-03-30)
+
+**Note:** Version bump only for package inferno-particles-demo
+
+
+
+
+
 # [2.0.0-alpha.5](https://github.com/matteobruni/tsparticles/compare/inferno-particles-demo@1.3.0...inferno-particles-demo@2.0.0-alpha.5) (2021-03-14)
 
 **Note:** Version bump only for package inferno-particles-demo