+++ demo/jquery/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.0.0-alpha.6](https://github.com/matteobruni/tsparticles/compare/jquery-particles-demo@1.9.2...jquery-particles-demo@2.0.0-alpha.6) (2021-03-30)
+
+
+### Features
+
+* added noise generator management like shape drawers and plugins ([0ae5ac6](https://github.com/matteobruni/tsparticles/commit/0ae5ac620c43199557f56c76517efbc7c178bd29))
+
+
+
+
+
 # [2.0.0-alpha.5](https://github.com/matteobruni/tsparticles/compare/jquery-particles-demo@1.5.0...jquery-particles-demo@2.0.0-alpha.5) (2021-03-14)