+++ demo/vanilla/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.0.0-alpha.6](https://github.com/matteobruni/tsparticles/compare/tsparticles-demo@1.9.2...tsparticles-demo@2.0.0-alpha.6) (2021-03-30)
+
+
+### Features
+
+* speed decay ([4c09893](https://github.com/matteobruni/tsparticles/commit/4c09893c6fc0105342a19c0424ec0aa9db338e3b))
+
+
+
+
+
 # [2.0.0-alpha.5](https://github.com/matteobruni/tsparticles/compare/tsparticles-demo@1.5.0...tsparticles-demo@2.0.0-alpha.5) (2021-03-14)
 
 **Note:** Version bump only for package tsparticles-demo