+++ demo/vue/CHANGELOG.md
@@ -3,6 +3,23 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.0.0-alpha.6](https://github.com/matteobruni/tsparticles/compare/particles.vue-demo@1.9.2...particles.vue-demo@2.0.0-alpha.6) (2021-03-30)
+
+
+### Bug Fixes
+
+* fixed vue plugin, thanks to [@tony19](https://github.com/tony19) ([e4b27bf](https://github.com/matteobruni/tsparticles/commit/e4b27bf6c7c8ef8d72844db80f24bc8be9da3380))
+* orbit color is now a string/color and not a hsl object ([fd5df21](https://github.com/matteobruni/tsparticles/commit/fd5df21d4a2ba785a4f256f54b9c351c8b6af02e))
+
+
+### Features
+
+* ported confetti to v2 architecture ([5aad60a](https://github.com/matteobruni/tsparticles/commit/5aad60ab474907f404316a9613d1a5a5605f8e32))
+
+
+
+
+
 # [2.0.0-alpha.5](https://github.com/matteobruni/tsparticles/compare/particles.vue-demo@1.5.0...particles.vue-demo@2.0.0-alpha.5) (2021-03-14)