+++ engine/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.0.0-alpha.6](https://github.com/matteobruni/tsparticles/compare/tsparticles-engine@2.0.0-alpha.5...tsparticles-engine@2.0.0-alpha.6) (2021-03-30)
+
+
+### Bug Fixes
+
+* fixed decay default value ([191f51a](https://github.com/matteobruni/tsparticles/commit/191f51a126735a9cc01acd9a8a9c7fa11b110d48))
+
+
+### Features
+
+* speed decay ([4c09893](https://github.com/matteobruni/tsparticles/commit/4c09893c6fc0105342a19c0424ec0aa9db338e3b))
+
+
+
+
+
 # [2.0.0-alpha.5](https://github.com/matteobruni/tsparticles/compare/tsparticles-engine@2.0.0-alpha.4...tsparticles-engine@2.0.0-alpha.5) (2021-03-14)
 
 **Note:** Version bump only for package tsparticles-engine