+++ presets/bubbles/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.0.0-alpha.6](https://github.com/matteobruni/tsparticles/compare/tsparticles-preset-bubbles@1.9.2...tsparticles-preset-bubbles@2.0.0-alpha.6) (2021-03-30)
+
+
+### Features
+
+* ported confetti to v2 architecture ([5aad60a](https://github.com/matteobruni/tsparticles/commit/5aad60ab474907f404316a9613d1a5a5605f8e32))
+
+
+
+
+
 # [2.0.0-alpha.5](https://github.com/matteobruni/tsparticles/compare/tsparticles-preset-bubbles@1.5.0...tsparticles-preset-bubbles@2.0.0-alpha.5) (2021-03-14)
 
 **Note:** Version bump only for package tsparticles-preset-bubbles